requirements.txt
flask>=3.0
psycopg2-binary>=2.9
gunicorn>=21.0
python-dotenv>=1.0
bcrypt>=4.0
weasyprint>=60.0
